lcw
sync

+++ TiJian/TiJian/ViewControllers/个人中心/收货地址/AddAddressController.m
@@ -95,7 +95,7 @@
                 _selectProvinceId = [pro_id integerValue];
                 _selectCityId = [city_id integerValue];
                 
-                NSLog(@"\nproId:%ld proName:%@\n cityId:%ld cityName:%@",self.provinceId,self.provinceName,self.cityId,self.cityName);
+                NSLog(@"\nproId:%ld proName:%@\n cityId:%ld cityName:%@",(long)self.provinceId,self.provinceName,self.cityId,self.cityName);
                 
             }else if (i == 3){
                 tf.text = self.addressModel.street;
@@ -205,7 +205,7 @@
     NSDictionary *params;
     NSString *api;
     
-    NSLog(@"proId:%ld proName:%@\n cityId:%ld cityName:%@",self.provinceId,self.provinceName,self.cityId,self.cityName);
+    NSLog(@"proId:%ld proName:%@\n cityId:%ld cityName:%@",(long)self.provinceId,self.provinceName,self.cityId,self.cityName);
     
     //编辑
     if (self.isEditAddress) {
@@ -255,7 +255,7 @@
 - (void)alertView:(UIAlertView *)alertView clickedButtonAtIndex:(NSInteger)buttonIndex
 {
     
-    NSLog(@"alertView proId:%ld proName:%@\n cityId:%ld cityName:%@",self.provinceId,self.provinceName,self.cityId,self.cityName);
+    NSLog(@"alertView proId:%ld proName:%@\n cityId:%ld cityName:%@",(long)self.provinceId,self.provinceName,self.cityId,self.cityName);
     
     if (buttonIndex == 1) {
         
@@ -276,7 +276,7 @@
     if ([self allTextFieldIsOK]) {
         //需要保存
         
-        NSLog(@"proId:%ld proName:%@\n cityId:%ld cityName:%@",self.provinceId,self.provinceName,self.cityId,self.cityName);
+        NSLog(@"proId:%ld proName:%@\n cityId:%ld cityName:%@",(long)self.provinceId,self.provinceName,self.cityId,self.cityName);
         UIAlertView *alert = [[UIAlertView alloc]initWithTitle:nil message:@"是否保存当前编辑信息" delegate:self cancelButtonTitle:@"否" otherButtonTitles:@"是", nil];
         [alert show];
         
@@ -498,7 +498,7 @@
     _selectProvinceId = self.provinceId;
     _selectCityId = self.cityId;
     
-    NSLog(@"在这里  省:%@ id %ld   市:%@ id:%ld",self.provinceName,self.provinceId,self.cityName,self.cityId);
+    NSLog(@"在这里  省:%@ id %ld   市:%@ id:%ld",self.provinceName,(long)self.provinceId,self.cityName,self.cityId);
     [self textFieldForTag:102].text = [NSString stringWithFormat:@"%@%@",self.provinceName,self.cityName];
 }
